# spec/api/openapi.yaml
openapi: "3.0.3"
info:
  title: AI Chat Orchestrator API
  description: |
    API สำหรับระบบ AI Portal กลางภาครัฐ — รับคำถามจากประชาชน วิเคราะห์ intent
    ค้นหาหน่วยงานที่เกี่ยวข้องแบบ runtime ผ่าน MCP Catalog และสังเคราะห์คำตอบ Markdown

    ## Flow การทำงาน (2 LLM calls per request)
    1. รับ `query` + `mcp_endpoint_url` จาก client
    2. ดึง Catalog หน่วยงานจาก MCP URL แบบ runtime
    3. LLM analyze_query (CoT): classify intent + decompose + route + label ในครั้งเดียว
    4. เรียก Agency APIs แบบขนาน พร้อม timeout + retry + circuit breaker
    5. LLM verify ความเกี่ยวข้องต่อ agency (threshold <= 0.5 → reject)
    6. v1: LLM สังเคราะห์คำตอบ Markdown / v2: raw answers / v3: sections + debug
  version: "2.0.0"
  contact:
    name: AI Portal Team

servers:
  - url: http://localhost:8000
    description: Local Development

paths:
  /v1/chat:
    post:
      operationId: chat
      summary: ส่งคำถามและรับคำตอบสังเคราะห์จากหลายหน่วยงาน
      description: |
        Endpoint หลักของ orchestrator — รับคำถามจากประชาชนและสังเคราะห์คำตอบจากหลายหน่วยงาน
        คำถาม chitchat จะได้รับคำตอบตรงจาก LLM โดยไม่เรียก agency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
            examples:
              search:
                summary: คำถามที่ต้องค้นข้อมูลหน่วยงาน
                value:
                  query: "ยาพาราเซตามอลต้องขึ้นทะเบียนอย่างไร"
                  mcp_endpoint_url: "https://mcp.example.com/catalog"
              chitchat:
                summary: คำถาม chitchat ทั่วไป
                value:
                  query: "สวัสดี"
                  mcp_endpoint_url: "https://mcp.example.com/catalog"

      responses:
        "200":
          description: ประมวลผลสำเร็จ (รวม case ที่หน่วยงานทั้งหมด fail — ดู errors field)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
              examples:
                search_success:
                  summary: ค้นหาสำเร็จ
                  value:
                    data:
                      answer: "### การขึ้นทะเบียนยา\nต้องยื่นเอกสาร..."
                      agencies:
                        - id: "fda"
                          name: "สำนักงาน อย."
                          relevance_score: 0.95
                          query: "ขั้นตอนการขึ้นทะเบียนยาพาราเซตามอล"
                      errors: []
                    meta:
                      intent: "search"
                      synthesized: true
                      responseTimeMs: 3420
                chitchat:
                  summary: chitchat ตอบตรง
                  value:
                    data:
                      answer: "สวัสดีครับ มีอะไรให้ช่วยไหม?"
                      agencies: []
                      errors: []
                    meta:
                      intent: "chitchat"
                      synthesized: false
                      responseTimeMs: 620
                all_failed:
                  summary: หน่วยงานทั้งหมด fail
                  value:
                    data:
                      answer: "ไม่สามารถดึงข้อมูลจากหน่วยงานได้ในขณะนี้"
                      agencies: []
                      errors:
                        - agency: "fda"
                          errorType: "Timeout Error"
                          message: "Connection timed out after 10000ms"
                    meta:
                      intent: "search"
                      synthesized: false
                      responseTimeMs: 10150
        "400":
          description: Request ไม่ถูกต้อง
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Missing query parameter"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v2/chat:
    post:
      operationId: chatV2
      summary: Raw agency answers ไม่มี LLM synthesis
      description: คำตอบต้นฉบับจากแต่ละหน่วยงานแยกกัน ไม่มีการสังเคราะห์
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: ประมวลผลสำเร็จ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatV2Response"
        "400":
          description: Request ไม่ถูกต้อง
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v3/chat:
    post:
      operationId: chatV3
      summary: Sections per sub-question + debug struct
      description: |
        แยกคำตอบต่อ sub-question ไม่มี LLM synthesis — `meta.synthesized` เป็น `false` เสมอ
        `data.debug` มีรายละเอียด pipeline ครบ (intent, timing, per-agency score, routing)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: ประมวลผลสำเร็จ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatV3Response"
        "400":
          description: Request ไม่ถูกต้อง
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /health:
    get:
      operationId: healthCheck
      summary: Health check
      responses:
        "200":
          description: Service is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"

components:
  schemas:
    # ──────────────── Request ────────────────
    ChatRequest:
      type: object
      required: [query, mcp_endpoint_url]
      additionalProperties: false
      properties:
        query:
          type: string
          description: คำถามจากผู้ใช้ (whitespace จะถูก normalize อัตโนมัติ)
          example: "ยาพาราเซตามอลต้องขึ้นทะเบียนอย่างไร"
        mcp_endpoint_url:
          type: string
          format: uri
          description: HTTPS URL ของ MCP Catalog สำหรับ discover หน่วยงานแบบ runtime
          example: "https://mcp.example.com/catalog"
        session_id:
          type: string
          nullable: true
          description: Session ID สำหรับ tracking ฝั่ง client (ถ้าไม่ส่งมา server จะ generate ให้)
          example: "550e8400-e29b-41d4-a716-446655440000"

    # ──────────────── Response ────────────────
    ChatResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/ChatData"
        meta:
          $ref: "#/components/schemas/ResponseMeta"

    ChatData:
      type: object
      required: [answer, agencies, errors]
      properties:
        answer:
          type: string
          description: คำตอบ Markdown หรือข้อความ fallback
          example: "### การขึ้นทะเบียนยา\nต้องยื่นเอกสาร..."
        agencies:
          type: array
          description: หน่วยงานที่ถูก query สำเร็จ
          items:
            $ref: "#/components/schemas/AgencySummary"
        errors:
          type: array
          description: รายละเอียด error จากหน่วยงานที่ fail
          items:
            $ref: "#/components/schemas/AgencyErrorDetail"

    ChatV2Response:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [agencies, errors]
          properties:
            agencies:
              type: array
              description: คำตอบต้นฉบับจากแต่ละหน่วยงาน (รวมถึงที่ relevance ต่ำ)
              items:
                type: object
                properties:
                  id: { type: string }
                  name: { type: string, nullable: true }
                  query: { type: string, nullable: true }
                  answer: { type: string }
                  relevance_score: { type: number, nullable: true, minimum: 0.0, maximum: 1.0 }
            errors:
              type: array
              items:
                $ref: "#/components/schemas/AgencyErrorDetail"
        meta:
          $ref: "#/components/schemas/ResponseMeta"

    ChatV3Response:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [answer, sections, errors]
          properties:
            answer:
              type: string
              description: Raw answers concat ด้วย ## topic headers (ไม่ผ่าน LLM)
            sections:
              type: array
              items:
                type: object
                properties:
                  title: { type: string }
                  agencies:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string, nullable: true }
                        query: { type: string, nullable: true }
                        content: { type: string }
            errors:
              type: array
              items:
                $ref: "#/components/schemas/AgencyErrorDetail"
            debug:
              type: object
              nullable: true
              description: Pipeline debug info — intent, timing, per-agency score, routing
        meta:
          $ref: "#/components/schemas/ResponseMeta"

    ResponseMeta:
      type: object
      required: [intent, synthesized, responseTimeMs, session_id]
      properties:
        intent:
          type: string
          enum: [search, chitchat, capability]
          description: ประเภท intent ที่ LLM วิเคราะห์ได้
          example: "search"
        synthesized:
          type: boolean
          description: true = LLM สังเคราะห์จาก agencies / false = chitchat ตอบตรง
          example: true
        responseTimeMs:
          type: integer
          minimum: 0
          description: เวลาประมวลผลรวม (milliseconds)
          example: 3420
        session_id:
          type: string
          description: Echo กลับจาก request หรือ UUID ที่ server generate ถ้าไม่ส่งมา
          example: "550e8400-e29b-41d4-a716-446655440000"

    AgencySummary:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          description: Agency ID จาก MCP Catalog
          example: "fda"
        name:
          type: string
          description: ชื่อหน่วยงาน
          example: "สำนักงาน อย."
        relevance_score:
          type: number
          format: float
          nullable: true
          minimum: 0.0
          maximum: 1.0
          description: คะแนนความเกี่ยวข้องของคำตอบ 0.0–1.0
          example: 0.95
        query:
          type: string
          nullable: true
          description: sub-question ที่ส่งไปยังหน่วยงานนี้
          example: "ขั้นตอนการขึ้นทะเบียนยาพาราเซตามอล"

    AgencyErrorDetail:
      type: object
      required: [agency, errorType, message]
      properties:
        agency:
          type: string
          description: Agency ID ที่ fail
          example: "revenue"
        errorType:
          type: string
          description: ประเภท error เช่น TimeoutError, MCPDiscoveryError, LLMGenerationError
          example: "TimeoutError"
        message:
          type: string
          description: รายละเอียด error
          example: "Connection timed out after 10000ms"

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Missing query parameter"